+++ NEWS.md
@@ -1,3 +1,11 @@
+# meditate 0.1.3
+
+- In `Meditate`, add new arguments `mandala` and `user_stops`.
+
+- In `ReadSessions`, fix bug with current streak not reporting correctly.
+
+- Add `PlotMandala`, draws a colored mandala.
+
 # meditate 0.1.2
 
-- First version of **meditate**
+- Initial CRAN release.